Misc fixes + constructor visibilities

+++ contracts/contracts/buyback/Buyback.sol
@@ -58,7 +58,7 @@ contract Buyback is Governable {
         if (uniswapAddr == address(0)) return;
         // Give Uniswap unlimited OUSD allowance
         ousd.safeApprove(uniswapAddr, 0);
-        ousd.safeApprove(uniswapAddr, type(uint).max);
+        ousd.safeApprove(uniswapAddr, type(uint256).max);
         emit UniswapUpdated(_address);
     }
